requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
streamlit==1.39.0
openai==0.28.0
python-dotenv==1.0.0